build_deps: Clean up the packages we install as build dependencies
A number of packages have crept in over the years, and they aren't
needed for the modern scripts, or they were added because things were
planned that never worked out.  For now, clean it up to things we only
use.  We can always add them back if a script uses it, or if a user
submits a community contributed script and a dependency is needed for
them.