c/functions: Pad images to 512-byte and quoting changes
We haven't been padding the images to the 512-byte sector boundary and
losetup always complained about it.  The RaspberryPi imager will also
refuse to write them as custom images because of this.  This is really
noticed with the slim images, so, lets finally fix it up and pad the
images.