rpi*: Update to enable cloud-init
This enables cloud-init, so if the user enables customization options in
the RaspberryPi Imager, aside from the network-config file bug
(upstream, not ours), they work.  We leave our other services enabled,
for the cases where the user does NOT choose to enable customization
options, they end up with the kali/kali defaults still.