bin/post-release: Set init_format to cloudinit
THe init_format specifier does not actually mean what init system the
images use, but instead specifies what the RPi Imager software should
use for setting up customizations.  Now that we have the changes needed
merged in, change it so that the rpi-imager.json will also generate it
and images will get the proper setup when using the rpi imager software
with releases.