bsp/cloudinit: Skip apt setup
We set it up a different way in the scripts, so remove these options.
Additionally, because we do not specify any in our config, currently it
is hit or miss in if it ignores the apt repos, or it sets them to both
ubuntu and debian repos (that do not exist) and not at all kali repos.

Thanks to S3L333 for reporting.