c/base_image: Ensure charmap is set to utf8
In d893b90a0892 we set the console-setup settings back to their
defaults, however, the actual issue with the prompt on ttys appears to
be due to the lack of utf8 setting in the charmap.

This commit is essentially a revert of d893b90a0892 and setting the
charmap to be utf8.