raspberry*: Move to using packaged kernel and firmware
We are now packaging RaspberryPi OS version of raspi-firmware, and the
6.6 kernel in the kali package repository, as well, the nexmon driver is
now a dkms module that can be installed from there as well.

Switching to these requires changes to where we mount the first
partition.  Additionally, on the arm64 images, we install both 2712 and
v8 kernels.  On armhf images, we install both v7 and v7l kernels, and
for the armel images we install -v6.

The services are modified to look on the new mount point for their
files.  The config.txt is greatly simplified because of the newer
firmware, and we include a config.txt in /boot that warns the user that
they should not actually use it, and point them to the correct location.