Initial work towards switching back to using chroot
systemd-nspawn can not be run easily under docker/in CI, so we are going
to move back to using chroot.  The original reason for switching away
from chroot was mount handling and on some devices, the rootfs would
corrupt itself when resizing.

Currently there seems to be no corruption when testing images on this
branch, but further testing across the board is required.