common: functions: Rework logging to remove ANSI from the log files
Because we are using colors on the console to make the output easier to
quickly scan through to find issues, when logging is enabled, we would
end up with ANSI codes in the logs which made them harder to read.
Rework this so that we still get colors on the console, and strip out
ANSI codes from the log file to make it easier to read.