c/functions: Set TERM to xterm-256color
When the scripts were first converted in 9827596dec, we never actually
tested setting TERM to something invalid, or that does not exist in
terminfo (because who would do such a thing?)

As it turns out, fakemachine sets the TERM to be vt220, and that does
not seem to exist in terminfo, so when tput gets run, it checks the
terminfo database, doesn't find it, and errors out with no output to
show the cause.

Tracked down by running the script with `strace -f` in order to follow
every process and find the crashing one. We can fix this a couple of
different ways, switching to use ANSI codes for colors, or by simply
setting the terminal to something that does exist in the terminfo, in
this case, `xterm-256color` will be used.