newport: Use truncate to create the image file
Prior to this commit, an intermediary image was being used in order to
create the image file, and this causes the temporary build size to
baloon up to needing around 61GB of space.  By using truncate instead,
we can increase the size of the image file instead, saving space during
the build.