common: fix limit_cpu to not swallow result
Occasionally we see the following when attempting to compress images

 Error writing block data
 ./common.d/functions.sh: line 632: pop_var_context: head of shell_variables not a function context
 common.d/finish_image.sh: line 48: pop_var_context: head of shell_variables not a function context

This error comes from pixz, but unfortunately, we were swallowing the
result code in limit_cpu, so while the compression would fail, limit_cpu
would succeed, thus continuing the script instead of bailing with an
error.  With this commit, we no longer swallow the result code.